+++ packages/config-plugins/src/plugins/expo-plugins.ts
@@ -6,8 +6,6 @@ import { ExpoConfig } from '@expo/config-types';
 import { ConfigPlugin, StaticPlugin } from '../Plugin.types';
 import * as AndroidConfig from '../android';
 import * as IOSConfig from '../ios';
-import { withPlugins } from './core-plugins';
-import { withStaticPlugin } from './static-plugins';
 import withAdMob from './unversioned/expo-ads-admob';
 import withAppleAuthentication from './unversioned/expo-apple-authentication';
 import withBranch from './unversioned/expo-branch';
@@ -17,6 +15,8 @@ import withNotifications from './unversioned/expo-notifications';
 import withSplashScreen from './unversioned/expo-splash-screen';
 import withUpdates from './unversioned/expo-updates';
 import withMaps from './unversioned/react-native-maps';
+import { withPlugins } from './withPlugins';
+import { withStaticPlugin } from './withStaticPlugin';
 
 /**
  * Config plugin to apply all of the custom Expo iOS config plugins we support by default.
