+++ packages/config-plugins/src/plugins/withRunOnce.ts
@@ -1,27 +1,5 @@
-import assert from 'assert';
-
-import { ConfigPlugin, StaticPlugin } from '../Plugin.types';
+import { ConfigPlugin } from '../Plugin.types';
 import { addHistoryItem, getHistoryItem, PluginHistoryItem } from '../utils/history';
-import { withStaticPlugin } from './static-plugins';
-
-/**
- * Resolves a list of plugins.
- *
- * @param config exported config
- * @param plugins list of config config plugins to apply to the exported config
- */
-export const withPlugins: ConfigPlugin<(StaticPlugin | ConfigPlugin | string)[]> = (
-  config,
-  plugins
-) => {
-  assert(
-    Array.isArray(plugins),
-    'withPlugins expected a valid array of plugins or plugin module paths'
-  );
-  return plugins.reduce((prev, plugin) => {
-    return withStaticPlugin(prev, { plugin });
-  }, config);
-};
 
 /**
  * Prevents the same plugin from being run twice.
