+++ packages/config-plugins/src/plugins/withStaticPlugin.ts
